requirements.txt
openai-whisper==20231117
torch>=2.0.0
torchaudio>=2.0.0
numpy>=1.21.0
librosa>=0.10.0
soundfile>=0.12.0
ffmpeg-python>=0.2.0
streamlit>=1.28.0
gradio>=4.0.0
python-dotenv>=1.0.0
requests>=2.31.0 